requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.0.3
django-model-utils==4.2.0
django-notifications-hq==1.7.0
idna==3.4
joblib==0.14.1
jsonfield==3.1.0
numpy==1.23.3
panda==0.3.1
pandas==1.5.0
Pillow==9.2.0
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
scipy==1.9.1
six==1.16.0
sqlparse==0.4.3
swapper==1.3.0
urllib3==1.26.12
